Converted and moved the example file into a test file and removed the README.md changes specific to the repo fork.
